Older Build/requirements.txt
fastapi
uvicorn
jinja2
azure-storage-blob
httpx
python-multipart
gunicorn
azure-identity
azure-keyvault-secrets
